// fronend/src/app/core/models/role.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMIN = 'ADMIN',
  BRANCH_MANAGER = 'BRANCH_MANAGER',
  ACCOUNTANT = 'ACCOUNTANT',
  CASHIER = 'CASHIER',
  LOAN_OFFICER = 'LOAN_OFFICER',
  CUSTOMER_SERVICE = 'CUSTOMER_SERVICE',
  ATM_MANAGER = 'ATM_MANAGER',
  AUDITOR = 'AUDITOR',
  CUSTOMER = 'CUSTOMER',
}